backend/requirements/perf.txt
pytest-benchmark==4.0.0
locust==2.27.0 